Move duplicate code to functions
